Add Authorization Checks

+++ Rest API - Social Media Backend/controllers/feed.js
@@ -138,11 +138,11 @@ exports.updatePost = (req, res, next) => {
 				error.statusCode = 404;
 				throw error;
 			}
-			// if (post.creator.id !== req.userId) {
-			// 	const error = new Error('Not authorized.');
-			// 	error.statusCode = 403;
-			// 	throw error;
-			// }
+			if (post.creator.toString() !== req.userId) {
+				const error = new Error('Not authorized.');
+				error.statusCode = 403;
+				throw error;
+			}
 			if (imageUrl !== post.imageUrl) {
 				clearImage(post.imageUrl);
 			}
@@ -180,11 +180,11 @@ exports.deletePost = (req, res, next) => {
 				error.statusCode = 404;
 				throw error;
 			}
-			// if (post.creator.id !== req.userId) {
-			//     const error = new Error('Not authorized.');
-			//     error.statusCode = 403;
-			//     throw error;
-			// }
+			if (post.creator.toString() !== req.userId) {
+				const error = new Error('Not authorized.');
+				error.statusCode = 403;
+				throw error;
+			}
 			clearImage(post.imageUrl);
 			return Post.findByIdAndRemove(postId);
 		})
